+++ src/random_search.rs
@@ -4,7 +4,7 @@ use std::{
     f64,
     ops::Range,
     sync::{
-        atomic::{AtomicBool, AtomicUsize, Ordering},
+        atomic::{AtomicBool, AtomicU32, Ordering},
         Arc, Mutex,
     },
     thread,
@@ -33,14 +33,14 @@ pub fn random_search<
     T: 'static + Copy + Send + Sync + Default + SampleUniform + PartialOrd,
     const N: usize,
 >(
-    iterations: usize,
+    iterations: u32,
     ranges: [Range<T>; N],
     f: fn(&[T; N]) -> f64,
     polling: Option<u64>,
     early_exit_minimum: Option<f64>,
 ) -> [T; N] {
     // Gets cpu data
-    let cpus = num_cpus::get();
+    let cpus = num_cpus::get() as u32;
     let remainder = iterations % cpus;
     let per = iterations / cpus;
 
@@ -50,18 +50,18 @@ pub fn random_search<
         remainder,
         ranges_arc.clone(),
         f,
-        // Since we aare doing this on the same thread, we don't need to use these
-        Arc::new(AtomicUsize::new(Default::default())),
+        // Since we are doing this on the same thread, we don't need to use these
+        Arc::new(AtomicU32::new(Default::default())),
         Arc::new(Mutex::new(Default::default())),
         Arc::new(AtomicBool::new(false)),
     );
 
     let thread_exit = Arc::new(AtomicBool::new(false));
     // (handles,(counters,thread_bests))
-    let (handles, links): (Vec<_>, Vec<(Arc<AtomicUsize>, Arc<Mutex<f64>>)>) = (0..cpus)
+    let (handles, links): (Vec<_>, Vec<(Arc<AtomicU32>, Arc<Mutex<f64>>)>) = (0..cpus)
         .map(|_| {
             let ranges_clone = ranges_arc.clone();
-            let counter = Arc::new(AtomicUsize::new(0));
+            let counter = Arc::new(AtomicU32::new(0));
             let thread_best = Arc::new(Mutex::new(f64::MAX));
 
             let counter_clone = counter.clone();
@@ -82,7 +82,7 @@ pub fn random_search<
             )
         })
         .unzip();
-    let (counters, thread_bests): (Vec<Arc<AtomicUsize>>, Vec<Arc<Mutex<f64>>>) =
+    let (counters, thread_bests): (Vec<Arc<AtomicU32>>, Vec<Arc<Mutex<f64>>>) =
         links.into_iter().unzip();
 
     if let Some(poll_rate) = polling {
@@ -115,10 +115,10 @@ pub fn random_search<
         T: 'static + Copy + Send + Sync + Default + SampleUniform + PartialOrd,
         const N: usize,
     >(
-        iterations: usize,
+        iterations: u32,
         ranges: Arc<[Range<T>; N]>,
         f: fn(&[T; N]) -> f64,
-        counter: Arc<AtomicUsize>,
+        counter: Arc<AtomicU32>,
         best: Arc<Mutex<f64>>,
         thread_exit: Arc<AtomicBool>,
     ) -> (f64, [T; N]) {
